ClauseAI_groq/backend/requirements.txt
Flask
flask-cors
groq
python-dotenv
PyPDF2
python-docx
